+++ R/randolist.R
@@ -8,6 +8,10 @@
 #' @param strata named list of stratification variables (see examples)
 #' @param blocksizes numbers of each arm to include in blocks (see details)
 #' @param pascal logical, whether to use pascal's triangle to determine block sizes
+#' @param n_init number of blocks with different blocksize probabilities to
+#' initialize the list with
+#' @param init_probs probabilities to use for each \code{blocksize} for the
+#' \code{n_init} blocks
 #' @param ... arguments passed on to other methods
 #'
 #' @details \code{blocksizes} defines the number of allocations to each arm in a block.
@@ -22,10 +26,28 @@
 #' intermediate sized blocks, which helps with making it more difficult to guess
 #' future allocations, and reduces the risk of finishing in the middle of a large
 #' block.
+#' If\code{pascal = FALSE}, all \code{blocksize}s have the same frequency.
 #'
 #' Unbalanced randomization is possible by specifying the same arm label multiple times.
 #'
 #' To disable block randomisation, set \code{blocksizes} to the same value as \code{n}.
+#'
+#' It can be helpful to use smaller blocks at the start of a randomisation list.
+#' \code{n_init} allows you to define how many blocks to add at the beginning of
+#' the list with a different set of blocksize probabilities (entered via
+#' \code{init_probs}). Note that this modifies the final frequencies of the
+#' blocksizes - they are no longer according to normal settings (pascals triangle, if
+#' \code{pascal = TRUE}, or approximately equal, if \code{pascal = FALSE}).
+#' Depending on the settings, it may even be that some blocksizes are not observed at
+#' all (e.g. if \code{n_init} exceeds the total number of blocks required and
+#' \code{init_probs = c(.8, .1, 0)}, there will be none of the larger blocks in
+#' the randomisation list.
+#'
+#' \code{n_init} and \code{init_probs} allow more control over the blocksize
+#' probabilities than is otherwise possible (i.e. with the \code{pascal} argument).
+#' Suppose you want primarily blocks of size 2, with some blocks of size 4, you might
+#' set \code{n_init} to \code{n/2} (as if all blocks were of size 2, just to ensure
+#' that there are enough blocks for all randomisations) and set \code{init_probs = c(0.8, .2)}.
 #'
 #' @export
 #'
@@ -49,13 +71,21 @@
 #' randolist(10, arms = c("arm 1", "arm 1", "arm 2"))
 #'
 #'
-randolist <- function(n, arms = LETTERS[1:2], strata = NA, blocksizes = 1:3, pascal = TRUE, ...){
+randolist <- function(n,
+                      arms = LETTERS[1:2],
+                      strata = NA,
+                      blocksizes = 1:3,
+                      pascal = TRUE,
+                      n_init = 0,
+                      init_probs = NULL,
+                      ...){
 
   strata_y <- is.list(strata)
 
   if(!strata_y) {
 
-    rlist <- blockrand(n = n, arms = arms, blocksizes = blocksizes, ...)
+    rlist <- blockrand(n = n, arms = arms, blocksizes = blocksizes, pascal = pascal,
+                       n_init = n_init, init_probs = init_probs, ...)
 
   } else {
 
@@ -68,7 +98,8 @@ randolist <- function(n, arms = LETTERS[1:2], strata = NA, blocksizes = 1:3, pas
       stratum <- nth[x]
 
       # generate randomization for this stratum
-      rlist <- blockrand(n = n, arms = arms, blocksizes = blocksizes, ...)
+      rlist <- blockrand(n = n, arms = arms, blocksizes = blocksizes, pascal = pascal,
+                         n_init = n_init, init_probs = init_probs, ...)
 
       # add stratum information to the randomization
       rlist$stratum <- stratum